Tests/LibWeb/Text/expected/HTML/displayed-image-survives-decoded-image-cache-pruning.txt

recreated image is completely available: false
